requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=1.10.0
python-dotenv>=1.0.0
httpx>=0.24.0
openai>=0.27.0
pytest>=7.0.0
requests>=2.28.0
google-generativeai>=0.1.0
